Fixed an ambiguous called to log() at least with some version of gcc

+++ rezound/src/backend/COSSSoundPlayer.cpp
@@ -58,7 +58,7 @@
 #define BUFFER_COUNT 4
 
 #define BUFFER_SIZE_BYTES (BUFFER_SIZE_FRAMES*sizeof(sample_t)*CHANNELS)	// buffer size in bytes
-#define BUFFER_SIZE_BYTES_LOG2 ((size_t)(log(BUFFER_SIZE_BYTES)/log(2.0)))	// log2(BUFFER_SIZE_BYTES) -- that is 2^this is BUFFER_SIZE_BYTES
+#define BUFFER_SIZE_BYTES_LOG2 ((size_t)(log((double)BUFFER_SIZE_BYTES)/log(2.0)))	// log2(BUFFER_SIZE_BYTES) -- that is 2^this is BUFFER_SIZE_BYTES
 
 
 COSSSoundPlayer::COSSSoundPlayer() :
